Update README.md
This motion-controlled Tetris game project brings a new twist to the classic game by incorporating motion control using an Arduino UNO R3 and various electronic components. The aim of this project is to provide an innovative and interactive gaming experience for users, allowing them to control the Tetris game using motion gestures and bringing a new level of immersion to the gameplay.